initial setup for basic vga writing project
i think i actually understand this now. I have some questions
that i need to ask an expert.
Is there a reason to not declare all wires at the top before
they are used? does this produce differnt results than just naming them
in module creation? is there a reason to not wrap all of these modules
up so that main only has to call one of them?